Move Homepage to components/Root and rename to index.jsx
