+++ frontend/src/App.jsx
@@ -1,8 +1,5 @@
 import './App.scss';
-
-import Button from './components/Button';
-import Bunny from './components/Bunny';
-import Homepage from './routes/Homepage';
+import Homepage from './components/Root';
 
 import { BrowserRouter, Routes, Route, Outlet } from 'react-router-dom';
